Extract vscode module tests to their own file
